Move session and constant config to MangoWebApplicationInitializer
